Build should work with CI and run all tests for MacOS
