+++ .travis.yml
@@ -11,10 +11,10 @@ before_install:
     - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update ; fi
     - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install llvm38 ; fi
     - export PHYRE_ROOT=/Users/travis/build/itsbriany/Phyre
+    - export PHYRE_RUNTIME=$PHYRE_ROOT/Build/Bin
 
 install:
-    - ./install.sh
+    - Tools/install.sh
 
-# Run the tests
 script:
-    - ./run_tests.sh
+    - Tools/run_tests.sh
