+++ Applications/GraphicsDemos/CMakeLists.txt
@@ -7,3 +7,4 @@ include_directories(${PHYRE_INCLUDE_DIRS})
 link_directories(${Boost_LIBRARY_DIRS} ${GLFW_LIBRARY_DIR})
 add_executable(DrawCube draw_cube.cpp draw_cube.h)
 target_link_libraries(DrawCube ${Boost_LIBRARIES}  ${VULKAN_LIBRARIES} glfw3 Logging Graphics)
+set_target_properties(DrawCube PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${PHYRE_RUNTIME})
