+++ Libs/Configuration/Test/configuration_tests.cpp
@@ -7,12 +7,14 @@ std::string path_to_phyre_config;
 
 class ConfigurationLoaderTest : public ::testing::Test {
 protected:
+    //-------------------- Setup & TearDown --------------------
     ConfigurationLoaderTest() : target_("ConfigurationLoaderTest") {
         p_loader_.reset(new Phyre::Configuration::Loader(path_to_phyre_config));
     }
 
     virtual ~ConfigurationLoaderTest() {}
 
+    //-------------------- Data Members ------------------------
     std::string target_;
     std::unique_ptr<Phyre::Configuration::Loader> p_loader_;
 };
@@ -26,7 +28,17 @@ TEST_F(ConfigurationLoaderTest, GetContents) {
     std::string resource("Example.txt");
     std::string contents = p_loader_->GetContents(target_, resource);
     EXPECT_FALSE(contents.empty());
+#if defined(_MSC_VER)
     EXPECT_STREQ("Hello\r\nworld", contents.c_str());
+#else
+    EXPECT_STREQ("Hello\nworld", contents.c_str());
+#endif
+}
+
+TEST_F(ConfigurationLoaderTest, GetContentsNoSuchResource) {
+    std::string resource("no_such_resource.res");
+    std::string contents = p_loader_->GetContents(target_, resource);
+    EXPECT_TRUE(contents.empty());
 }
 
 int main(int argc, char* argv[]) {
