+++ Libs/XMPP/CMakeLists.txt
@@ -2,12 +2,10 @@ cmake_minimum_required(VERSION 3.3)
 
 # Library
 include_directories(${PHYRE_INCLUDE_DIRS})
-add_library(XMPP xmpp_client.h xmpp_client.cpp xmpp_state.h sasl.h
-    sasl.cpp xmpp_tests.h)
+file(GLOB XMPP_SRC "*.c*" "*.h")
+add_library(XMPP ${XMPP_SRC})
 target_link_libraries(XMPP ${Boost_LIBRARIES} Networking b64)
+set_target_properties(XMPP PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${PHYRE_LIBS}
+                                      ARCHIVE_OUTPUT_DIRECTORY ${PHYRE_LIBS})
 
-
-include_directories(${PHYRE_INCLUDE_DIRS})
-link_directories(${Boost_LIBRARY_DIRS})
-add_executable(XMPPTests xmpp_tests.h xmpp_test_runner.cpp)
-target_link_libraries(XMPPTests ${Boost_LIBRARIES} gmock gtest XMPP)
+add_subdirectory(Test)
