+++ Libs/XMPP/Test/authentication_selection.xml
