+++ Libs/XMPP/Test/authentication_selection_chunk1.xml
