+++ Libs/XMPP/Test/authentication_selection_chunk2.xml
