+++ Libs/XMPP/Test/base64_challenge.xml
