+++ Libs/XMPP/Test/generic.xml
