--- Testing/.gitignore
@@ -1,1 +1,0 @@
-Bin
