--- Testing/CMakeLists.txt
@@ -1,7 +1,0 @@
-cmake_minimum_required(VERSION 3.3)
-if(Boost_FOUND)
-    include_directories(${PHYRE_INCLUDE_DIRS})
-    link_directories(${Boost_LIBRARY_DIRS})
-    add_executable(PhyreTests phyre_tests.cpp)
-    target_link_libraries(PhyreTests ${Boost_LIBRARIES} gmock Networking XMPP)
-endif()
